+++ zephyr/ringbuffer.c
@@ -21,7 +21,7 @@
 //  a pointer to a ringbuffer or NULL if there were no memory to allocate it.
 // Description: This function initializes the ring buffer.
 // dataSize is usually sizeof of some struct describing the data elements.
-ring_buffer_t *rb_init_ring_buffer(int size, size_t dataSize)
+ring_buffer_t *rb_init_ring_buffer(uint32_t size, uint32_t dataSize)
 {
     ring_buffer_t *cb = (ring_buffer_t *)malloc(sizeof(ring_buffer_t));
     if (cb == NULL) {
@@ -33,9 +33,9 @@ ring_buffer_t *rb_init_ring_buffer(int size, size_t dataSize)
         return NULL; // Memory allocation failed
     }
     cb->size = size;
-    cb->head = 0;
-    cb->tail = 0;
-    cb->count = 0;
+    cb->head = 0u;
+    cb->tail = 0u;
+    cb->count = 0u;
     cb->dataSize = dataSize;
     k_sem_init(&cb->bmx,1,1);
     k_sem_give(&cb->bmx);
@@ -151,7 +151,7 @@ void rb_scan(ring_buffer_t *cb, rb_scan_cb_t callback)
 {
     k_sem_take(&cb->bmx,K_FOREVER); // Lock the mutex
     // Initialize index to head
-    int currentIndex = cb->head / cb->dataSize;
+    uint32_t currentIndex = cb->head / cb->dataSize;
 
     // Scan all available data in the buffer
     for (uint32_t i = 0; i < cb->count; ++i) {
@@ -185,13 +185,13 @@ void* rb_inject(ring_buffer_t* cb, void* initial_value, rb_inject_cb_t callback)
     }
 
     // Initialize index to head
-    int currentIndex = cb->head / cb->dataSize;
+    uint32_t currentIndex = cb->head / cb->dataSize;
 
     // Initialize accumulated value to the initial value
     void *accumulatedValue = initial_value;
 
     // Perform the reduction operation on all elements in the buffer
-    for (int i = 0; i < cb->count; ++i) {
+    for (uint32_t i = 0; i < cb->count; ++i) {
         // Call the callback function with the current accumulated value and the current data element
         accumulatedValue = callback(accumulatedValue, cb->buffer + currentIndex * cb->dataSize);
 
@@ -214,7 +214,7 @@ void* rb_inject(ring_buffer_t* cb, void* initial_value, rb_inject_cb_t callback)
 // if creates new function from the original function (mapping).
 // The callback function takes two arguments: pointer to the curent original value and pointer to a space
 // where the new value must be stored.
-ring_buffer_t* rb_map(ring_buffer_t* cb, rb_map_cb_t callback, size_t mappedDataSize)
+ring_buffer_t* rb_map(ring_buffer_t* cb, rb_map_cb_t callback, uint32_t mappedDataSize)
 {
     k_sem_take(&cb->bmx,K_FOREVER); // Lock the mutex
 
@@ -234,7 +234,7 @@ ring_buffer_t* rb_map(ring_buffer_t* cb, rb_map_cb_t callback, size_t mappedData
     }
 
     // Apply the callback function to each element of the original circular buffer and enqueue the results into the new circular buffer
-    int currentIndex = cb->head / cb->dataSize;
+    uint32_t currentIndex = cb->head / cb->dataSize;
     for (uint32_t i = 0; i < cb->count; ++i) {
         callback(cb->buffer + currentIndex * cb->dataSize, new_data_element);
         rb_enqueue(rb_mapped,new_data_element);
